% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.readable.r
\docType{class}
\name{BinaryReadable}
\alias{BinaryReadable}
\title{Readable class}
\format{\code{\link{R6Class}} object.}
\usage{
BinaryReadable
}
\value{
Object of \code{\link{R6Class}} .
}
\description{
Readable class
}
\keyword{data}
