% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapackage.R
\name{datapackage}
\alias{datapackage}
\title{datapackage}
\usage{
datapackage(descriptor, base_path, strict=True)
}
\arguments{
\item{descriptor}{name}

\item{base_path}{json schema}

\item{strict}{descriptor}
}
\value{
...
}
\description{
...
}
\details{
...
}
\examples{
...
}
\seealso{
...
}
\author{
...
}
