% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPackageError.R
\docType{class}
\name{DataPackageError}
\alias{DataPackageError}
\title{DataPackageError class}
\format{\code{\link{R6Class}} object.}
\usage{
DataPackageError
}
\value{
Object of \code{\link{R6Class}} .
}
\description{
DataPackageError class
}
\keyword{data}
