% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dereferencePackageDescriptor}
\alias{dereferencePackageDescriptor}
\title{Dereference descriptor}
\usage{
dereferencePackageDescriptor(descriptor, basePath)
}
\arguments{
\item{descriptor}{descriptor}

\item{basePath}{basePath}
}
\description{
Dereference descriptor
}
