% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{descriptor.pointer}
\alias{descriptor.pointer}
\title{descriptor pointer}
\usage{
descriptor.pointer(value, baseDescriptor)
}
\arguments{
\item{value}{value}

\item{baseDescriptor}{baseDescriptor}
}
\description{
descriptor pointer
}
