% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{file_basename}
\alias{file_basename}
\title{File basename}
\usage{
file_basename(path)
}
\arguments{
\item{path}{character vector with path names}
}
\description{
Removes all of the path up to and including the last path separator (if any)  without extensions.
}
