% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{file_extension}
\alias{file_extension}
\title{File extension}
\usage{
file_extension(path)
}
\arguments{
\item{path}{string with path names}
}
\description{
Returns the file extension without the leading dot.
}
