% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{filterDefaultDialect}
\alias{filterDefaultDialect}
\title{Filter Default Dialect}
\usage{
filterDefaultDialect(dialect = NULL)
}
\arguments{
\item{dialect}{list}
}
\value{
dialect list
}
\description{
Helper function to filter default dialect
quoteChar and escapeChar are mutually exclusive: https://frictionlessdata.io/specs/csv-dialect/#specification
}
