% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{findFiles}
\alias{findFiles}
\title{findFiles}
\usage{
findFiles(pattern, path = getwd())
}
\arguments{
\item{pattern}{string pattern}

\item{path}{string path}
}
\description{
Find a file pattern in a specified directory.
}
