% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get.descriptor.path}
\alias{get.descriptor.path}
\title{get descriptor path}
\usage{
get.descriptor.path(directory= ".")
}
\arguments{
\item{directory}{A character vector of full path name. The default corresponds to the working directory specified by \code{\link[base]{getwd}}}
}
\description{
get descriptor path
}
