% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{infer}
\alias{infer}
\title{Infer a data package descriptor}
\usage{
infer(pattern = NULL, basePath = NULL)
}
\arguments{
\item{pattern}{string with file pattern}

\item{basePath}{base path for all relative paths}
}
\value{
Data Package Descriptor
}
\description{
A standalone function to infer a data package descriptor.
}
\examples{
\dontrun{
descriptor = infer("csv",basePath = '.')
descriptor
}

}
