% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is.compressed}
\alias{is.compressed}
\title{is compressed}
\usage{
is.compressed(x)
}
\arguments{
\item{x}{string with the file's path}
}
\value{
TRUE if file is compressed
}
\description{
Helper function to check if a file is compressed..
}
