% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is.local.descriptor.path}
\alias{is.local.descriptor.path}
\title{Is Local Descriptor Path}
\usage{
is.local.descriptor.path(descriptor, directory = ".")
}
\arguments{
\item{descriptor}{descriptor}

\item{directory}{A character vector of full path name. The default corresponds to the working directory specified by \code{\link[base]{getwd}}}
}
\description{
Helper functoin to check if a descriptor is local
}
