% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.valid.R
\name{is.valid}
\alias{is.valid}
\title{is.valid}
\usage{
is.valid(descriptor,schema)
}
\arguments{
\item{descriptor}{json dictionary}

\item{schema}{json schema}
}
\value{
...
}
\description{
...
}
\details{
...
}
\examples{
...
}
\seealso{
...
}
\author{
...
}
