% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{isRemotePath}
\alias{isRemotePath}
\title{Is remote path}
\usage{
isRemotePath(path)
}
\arguments{
\item{path}{string path}
}
\value{
TRUE if path is remote
}
\description{
Helper function to identify a remote path.
}
