% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{isSafePath}
\alias{isSafePath}
\title{Is safe path}
\usage{
isSafePath(path)
}
\arguments{
\item{path}{string path}
}
\value{
TRUE if path is safe
}
\description{
Helper function to check if a path is safe.
}
