% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{isUndefined}
\alias{isUndefined}
\title{Check if a variable is undefined or NULL}
\usage{
isUndefined(x)
}
\arguments{
\item{x}{variable}
}
\value{
TRUE if variable is undefined
}
\description{
Helper function to check if a variable is undefined or NULL.
}
