% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{class}
\name{Package}
\alias{Package}
\title{Package class}
\format{\code{\link{R6Class}} object}
\usage{
Package
}
\value{
Object of \code{\link{R6Class}}
}
\description{
Package class
}
\keyword{data}
