% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{Profile.load}
\alias{Profile.load}
\title{Instantiate \code{Profile} class}
\usage{
Profile.load(profile)
}
\arguments{
\item{profile}{string profile name in registry or URL to JSON Schema}
}
\value{
\code{\link{Profile}} class object
}
\description{
Constuctor to instantiate \code{\link{Profile}} class.
}
