% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{push}
\alias{push}
\title{Push elements in a list or vector}
\usage{
push(x, value)
}
\arguments{
\item{x}{list or vector}

\item{value}{value to push in x}
}
\description{
Push elements in a list or vector
}
