% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\docType{class}
\name{Resource}
\alias{Resource}
\title{Resource class}
\format{\code{\link{R6Class}} object.}
\usage{
Resource
}
\value{
Object of \code{\link{R6Class}} .
}
\description{
Resource class
}
\keyword{datasets}
