% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{retrieveDescriptor}
\alias{retrieveDescriptor}
\title{Retrieve descriptor}
\usage{
retrieveDescriptor(descriptor)
}
\arguments{
\item{descriptor}{descriptor}
}
\description{
Helper function to retrieve descriptor.
}
