% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{validate descriptor}
\usage{
validate(descriptor)
}
\arguments{
\item{descriptor}{data package descriptor, one of:
\itemize{
\item string with the local CSV file (path)
\item string with the remote CSV file (url)
\item list object
}}
}
\value{
A list with:
\itemize{
 \item{\code{valid }}{ \code{TRUE} if valid}  
 \item{\code{errors }}{a list with errors if valid \code{FALSE}}
 }
}
\description{
A standalone function to validate a data package descriptor.
}
\examples{
validate(descriptor = '{"name": "Invalid Datapackage"}')

}
\seealso{
\href{https://github.com/frictionlessdata/datapackage-r#validate}{https://github.com/frictionlessdata/datapackage-r#validate}
}
