% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.readable.connection.r
\docType{class}
\name{BinaryReadableConnection}
\alias{BinaryReadableConnection}
\title{Binary Readable Connection class}
\format{\code{\link{R6Class}} object.}
\usage{
BinaryReadableConnection
}
\value{
Object of \code{\link{R6Class}}.
}
\description{
Binary Readable connection class
}
\keyword{data}
