% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dereferenceResourceDescriptor}
\alias{dereferenceResourceDescriptor}
\title{Dereference resource descriptor}
\usage{
dereferenceResourceDescriptor(descriptor, basePath, baseDescriptor)
}
\arguments{
\item{descriptor}{descriptor}

\item{basePath}{basePath}

\item{baseDescriptor}{baseDescriptor}
}
\description{
Dereference resource descriptor
}
