% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{descriptor.pointer}
\alias{descriptor.pointer}
\title{Descriptor pointer}
\usage{
descriptor.pointer(value, baseDescriptor)
}
\arguments{
\item{value}{value  that specifies location in the descriptor}

\item{baseDescriptor}{base descriptor}
}
\description{
Helper function for specifying locations in a descriptor.
}
