% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{expandPackageDescriptor}
\alias{expandPackageDescriptor}
\title{Expand descriptor}
\usage{
expandPackageDescriptor(descriptor)
}
\arguments{
\item{descriptor}{descriptor}
}
\description{
Expand descriptor
}
