% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{expandResourceDescriptor}
\alias{expandResourceDescriptor}
\title{Expand resource descriptor}
\usage{
expandResourceDescriptor(descriptor)
}
\arguments{
\item{descriptor}{descriptor}
}
\description{
Helper function to expand resource descriptor.
}
