% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{helpers.from.list.to.json}
\alias{helpers.from.list.to.json}
\title{Convert list to json}
\usage{
helpers.from.list.to.json(json)
}
\arguments{
\item{json}{json string}
}
\description{
Convert list to json
}
