% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.valid.R
\name{is.valid}
\alias{is.valid}
\title{Is valid}
\usage{
is.valid(descriptor, schema = NULL)
}
\arguments{
\item{descriptor}{descriptor, one of:
\itemize{
\item string with the local CSV file (path)
\item string with the remote CSV file (url)
\item list object
}}

\item{schema}{Contents of the json schema, or a filename containing a schema}
}
\value{
\code{TRUE} if valid
}
\description{
Validate a descriptor over a schema
}
