% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\name{Package.load}
\alias{Package.load}
\title{Package.load}
\usage{
Package.load(descriptor = "{}", basePath = NULL, strict = FALSE)
}
\arguments{
\item{descriptor}{descriptor}

\item{basePath}{basePath}

\item{strict}{strict}
}
\description{
Package.load
}
