% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\docType{class}
\name{Profile}
\alias{Profile}
\title{Profile class}
\format{\code{\link{R6Class}} object.}
\usage{
Profile
}
\value{
Object of \code{\link{R6Class}} .
}
\description{
Profile class
}
\keyword{datasets}
