% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{resource}
\alias{resource}
\title{resource}
\usage{
resource(descriptor, base_path=None)
}
\arguments{
\item{descriptor}{descriptor}

\item{name}{name}

\item{jsonschema}{json schema}
}
\value{
...
}
\description{
...
}
\details{
...
}
\examples{
...
}
\seealso{
...
}
\author{
...
}
