% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{push}
\alias{push}
\title{Push elements in a list or vector}
\usage{
push(x, value)
}
\arguments{
\item{x}{list or vector}

\item{value}{object to push in x}
}
\description{
Helper function to add components in a list or vector.
}
