% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{Resource.load}
\alias{Resource.load}
\title{Resource.load}
\usage{
Resource.load(descriptor = list(), basePath = NA, strict = FALSE,
  dataPackage = list())
}
\arguments{
\item{descriptor}{descriptor}

\item{basePath}{basePath}

\item{strict}{strict}

\item{dataPackage}{dataPackage}
}
\description{
Resource.load
}
