% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{Resource.load}
\alias{Resource.load}
\title{Resource.load}
\usage{
Resource.load(descriptor = list(), basePath = NULL, strict = FALSE)
}
\arguments{
\item{descriptor}{descriptor}

\item{basePath}{basePath}

\item{strict}{strict}
}
\description{
Resource.load
}
